requirements.txt
matplotlib
numpy
seaborn
pandas
langchain 
langchain-community
langchain_groq
fastapi
uvicorn
python-dotenv
openpyxl
